[Kinetics] Put ReactionData implementations with matching rates
