+++ include/cantera/kinetics/ReactionData.h
@@ -94,18 +94,6 @@ protected:
 };
 
 
-//! Data container holding shared data specific to ArrheniusRate
-/**
- * The data container `ArrheniusData` holds precalculated data common to
- * all `ArrheniusRate` objects.
- */
-struct ArrheniusData : public ReactionData
-{
-    virtual bool update(const ThermoPhase& phase, const Kinetics& kin);
-    using ReactionData::update;
-};
-
-
 //! Data container holding shared data specific to TwoTempPlasmaRate
 /**
  * The data container `TwoTempPlasmaData` holds precalculated data common to
@@ -162,54 +150,6 @@ struct BlowersMaselData : public ReactionData
 
 protected:
     int m_state_mf_number; //!< integer that is incremented when composition changes
-};
-
-
-//! Data container holding shared data specific to Falloff rates
-/**
- * The data container `FalloffData` holds precalculated data common to
- * all Falloff related reaction rate classes.
- */
-struct FalloffData : public ReactionData
-{
-    FalloffData();
-
-    virtual bool update(const ThermoPhase& phase, const Kinetics& kin) override;
-
-    virtual void update(double T) override;
-
-    virtual void update(double T, double M) override;
-
-    using ReactionData::update;
-
-    //! Perturb third-body concentration vector of data container
-    /**
-     * The method is used for the evaluation of numerical derivatives.
-     * @param deltaM  relative third-body perturbation
-     */
-    void perturbThirdBodies(double deltaM);
-
-    virtual void restore() override;
-
-    virtual void resize(size_t nSpecies, size_t nReactions, size_t nPhases) override {
-        conc_3b.resize(nReactions, NAN);
-        m_conc_3b_buf.resize(nReactions, NAN);
-        ready = true;
-    }
-
-    virtual void invalidateCache() override {
-        ReactionData::invalidateCache();
-        molar_density = NAN;
-    }
-
-    bool ready; //!< boolean indicating whether vectors are accessible
-    double molar_density; //!< used to determine if updates are needed
-    vector_fp conc_3b; //!< vector of effective third-body concentrations
-
-protected:
-    int m_state_mf_number; //!< integer that is incremented when composition changes
-    bool m_perturbed; //!< boolean indicating whether 3-rd body values are perturbed
-    vector_fp m_conc_3b_buf; //!< buffered third-body concentrations
 };
 
 
@@ -296,48 +236,6 @@ struct ChebyshevData : public ReactionData
 
 protected:
     double m_pressure_buf; //!< buffered pressure
-};
-
-
-//! Data container holding shared data for reaction rate specification with interfaces
-/**
- * The data container InterfaceData holds precalculated data common to
- * InterfaceRate and StickingRate objects.
- *
- * The data container inherits from BlowersMaselData, where density is used to
- * hold the site density [kmol/m^2].
- */
-struct InterfaceData : public BlowersMaselData
-{
-    InterfaceData();
-
-    virtual bool update(const ThermoPhase& bulk, const Kinetics& kin) override;
-
-    virtual void update(double T) override;
-
-    virtual void update(double T, const vector_fp& values) override;
-
-    using BlowersMaselData::update;
-
-    virtual void perturbTemperature(double deltaT);
-
-    virtual void resize(size_t nSpecies, size_t nReactions, size_t nPhases) override {
-        coverages.resize(nSpecies, 0.);
-        logCoverages.resize(nSpecies, 0.);
-        partialMolarEnthalpies.resize(nSpecies, 0.);
-        electricPotentials.resize(nPhases, 0.);
-        standardChemPotentials.resize(nSpecies, 0.);
-        standardConcentrations.resize(nSpecies, 0.);
-        ready = true;
-    }
-
-    double sqrtT; //!< square root of temperature
-
-    vector_fp coverages; //!< surface coverages
-    vector_fp logCoverages; //!< logarithm of surface coverages
-    vector_fp electricPotentials; //!< electric potentials of phases
-    vector_fp standardChemPotentials; //!< standard state chemical potentials
-    vector_fp standardConcentrations; //!< standard state concentrations
 };
 
 }
